+++ magpie/api/management/service/service_formats.py
@@ -1,23 +1,30 @@
-from magpie.utils import get_twitcher_protected_service_url
-from magpie.services import service_type_dict
-from magpie.definitions.pyramid_definitions import *
 from magpie.api.api_except import evaluate_call
 from magpie.api.management.resource.resource_utils import crop_tree_with_permission
 from magpie.api.management.resource.resource_formats import get_resource_children, format_resource_tree
+from magpie.definitions.pyramid_definitions import *
+from magpie.definitions.sqlalchemy_definitions import Session
+from magpie.definitions.typedefs import Optional, JsonBody, AnyStr, Dict, List
+from magpie.utils import get_twitcher_protected_service_url
+from magpie.models import Resource, Service
+from magpie.services import service_type_dict, ServiceI
 
 
-def format_service(service, permissions=None, show_private_url=False):
+def format_service(service, permissions=None, show_private_url=False, show_resources_allowed=False):
+    # type: (Service, Optional[bool], Optional[bool], Optional[bool]) -> JsonBody
     def fmt_svc(svc, perms):
         svc_info = {
             u'public_url': str(get_twitcher_protected_service_url(svc.resource_name)),
             u'service_name': str(svc.resource_name),
             u'service_type': str(svc.type),
-            u'service_sync_type': str(svc.sync_type),
+            u'service_sync_type': str(svc.sync_type) if svc.sync_type is not None else svc.sync_type,
             u'resource_id': svc.resource_id,
             u'permission_names': sorted(service_type_dict[svc.type].permission_names if perms is None else perms)
         }
         if show_private_url:
             svc_info[u'service_url'] = str(svc.url)
+        if show_resources_allowed:
+            svc_info[u'resource_types_allowed'] = sorted(service_type_dict[svc.type].resource_types)
+            svc_info[u'resource_child_allowed'] = service_type_dict[svc.type].child_resource_allowed
         return svc_info
 
     return evaluate_call(
@@ -28,8 +35,13 @@ def format_service(service, permissions=None, show_private_url=False):
     )
 
 
-def format_service_resources(service, db_session, service_perms=None,
-                             resources_perms_dict=None, display_all=False, show_private_url=True):
+def format_service_resources(service,                       # type: Service
+                             db_session,                    # type: Session
+                             service_perms=None,            # type: Optional[List[AnyStr]]
+                             resources_perms_dict=None,     # type: Optional[Dict[AnyStr, List[AnyStr]]]
+                             display_all=False,             # type: Optional[bool]
+                             show_private_url=True,         # type: Optional[bool]
+                             ):                             # type: (...) -> JsonBody
     def fmt_svc_res(svc, db, svc_perms, res_perms, show_all):
         tree = get_resource_children(svc, db)
         if not show_all:
@@ -46,3 +58,12 @@ def format_service_resources(service, db_session, service_perms=None,
         msgOnFail="Failed to format service resources tree",
         content=format_service(service, service_perms, show_private_url=show_private_url)
     )
+
+
+def format_service_resource_type(resource_type, service_type):
+    # type: (Resource, ServiceI) -> JsonBody
+    return {
+        u'resource_type': resource_type.resource_type_name,
+        u'resource_child_allowed': resource_type.child_resource_allowed,
+        u'permission_names': service_type.resource_types_permissions[resource_type.resource_type_name],
+    }