+++ HISTORY.rst
@@ -15,7 +15,9 @@ History
 * use schema API route definitions for UI calls
 * use sub-requests API call for UI operations (fixes issue #114)
 * add new route `/services/types` to obtain a list of available service types
-* add `resources_allowed` and `resource_types_allowed` fields in service response
+* add `resource_child_allowed` and `resource_types_allowed` fields in service response
+* change service response for generic `service: {<info>}` instead of `{service-name}: {<info>}`
+* add new route `/services/types/{svc_type}/resources` for details about child service type resources
 * additional tests for new routes and operations previously left unevaluated
 
 0.8.x